Feladat02

+++ CheckoutScanner/src/main/java/org/example/Rule.java
@@ -15,7 +15,7 @@ public class Rule {
         return item;
     }
 
-    public int getCount() {
+    public double getCount() {
         return count;
     }
